% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{coalesce}
\alias{coalesce}
\alias{\%||\%}
\title{Coalesce}
\usage{
a \%||\% b
}
\arguments{
\item{a, b}{Returns first non-\code{NA}.}
}
\description{
Right side when left side is \code{NA}
}
\examples{
1 \%||\% 2 # returns 1
NA \%||\% 2 # returns 2
}
