% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{functions}
\alias{functions}
\title{Utility Functions}
\description{
Utility Functions
}
\section{Utility Functions}{
 \describe{
\item{\code{\link{make_names}}}{make_names (see \code{?"make_names"}).}
}
}

