% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_names}
\alias{make_names}
\title{Make names}
\usage{
make_names(x)
}
\arguments{
\item{x}{Vector of names.}
}
\description{
Analoguous to make.names, but replaces the dot with an underscore.
}
\examples{
make_names(c('FOO', 'bAR&', 'FOO BAR')) # returns c('foo', 'bar_', 'foo_bar')
}
