% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{operators}
\alias{operators}
\title{Infix Operators}
\description{
Infix Operators
}
\section{Infix Operators}{
 \describe{
\item{\code{\link{coalesce}}}{Coalesce (see \code{?"\%||\%"}).}
\item{\code{\link{paste0}}}{Concatenate (see \code{?"\%+\%"}).}
\item{\code{\link{paste.comma}}}{Concatenate comma separated (see \code{?"\%,\%"}).}
\item{\code{\link{paste.space}}}{Concatenate space separated (see \code{?"\% \%"}).}
\item{\code{\link{file.path}}}{Construct path (see \code{?"\%//\%"}).}
\item{\code{\link{not.in}}}{Negated in (see \code{?"\%!in\%"}).}
}
}

