% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{paste.space}
\alias{paste.space}
\alias{\% \%}
\alias{\%}
\title{Concatenate with space separation}
\usage{
x \% \% y
}
\arguments{
\item{x, y}{Vectors to be concatenated.}
}
\description{
Equivalent to \code{paste(x, y, sep = ' ')}, but removes \code{NA}.
}
\examples{
NA \% \% "this" \% \% 1900 # returns "this 1900"
}
