% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{paste0}
\alias{paste0}
\alias{\%+\%}
\title{Concatenate}
\usage{
x \%+\% y
}
\arguments{
\item{x, y}{Vectors to be concatenated.}
}
\description{
Equivalent to \code{paste0(x, y)}, but removes \code{NA}.
}
\examples{
NA \%+\% "this" \%+\% 1900 # returns "this1900"
}
